# Vectorize NewId formatters