--- a/tests/MassTransit.Abstractions.Tests/MassTransit.Abstractions.Tests.csproj
+++ b/tests/MassTransit.Abstractions.Tests/MassTransit.Abstractions.Tests.csproj
@@ -12,12 +12,10 @@
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.4.1" />
     <PackageReference Include="NUnit" Version="3.13.3" />
     <PackageReference Include="NUnit3TestAdapter" Version="4.3.1" />
+    <PackageReference Include="System.Text.Json" Version="6.0.2" />
   </ItemGroup>
 
   <ItemGroup>
-    <Content Include="NewId\guids.txt">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </Content>
     <Content Include="NewId\texts.txt">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </Content>