… tests/MassTransit.Abstractions.Tests/NewId/Formatter_Specs.cs
… tests/MassTransit.Abstractions.Tests/NewId/guids.txt
… tests/MassTransit.Abstractions.Tests/NewId/texts.txt