Mirror of [this NewId pr](https://github.com/phatboyg/NewId/pull/27).

- Vectorized all formatters including `ZBase32Formatter`.
- Removed allocations using stackalloc.
- Optimized to scalar `HexFormatter` using bit hacks.
- Added encoding unit tests.

### Benchmark
Code [found here](https://github.com/TimothyMakkison/NewId/blob/formatter_experiment_speed_test/tests/NewId.Benchmarks/Benchmarks.cs).

|               Method | EnvironmentVariables |      Mean |    Error |   StdDev | Ratio | Allocated |
|--------------------- |--------------------- |----------:|---------:|---------:|------:|----------:|
|      ToStringZBase32 | COMPlus_EnableSSE2=0 | 106.18 ns | 0.495 ns | 0.463 ns |  1.00 |      80 B |
|      ToStringZBase32 |                Empty |  56.51 ns | 0.907 ns | 0.758 ns |  0.53 |      80 B |
|                      |                      |           |          |          |       |           |
|       ToStringBase32 | COMPlus_EnableSSE2=0 | 108.60 ns | 0.441 ns | 0.368 ns |  1.00 |      80 B |
|       ToStringBase32 |                Empty |  60.27 ns | 0.960 ns | 0.801 ns |  0.55 |      80 B |
|                      |                      |           |          |          |       |           |
| ToStringCustomBase32 | COMPlus_EnableSSE2=0 | 106.70 ns | 0.390 ns | 0.346 ns |  1.00 |      80 B |
| ToStringCustomBase32 |                Empty |  75.59 ns | 0.808 ns | 0.716 ns |  0.71 |      80 B |
|                      |                      |           |          |          |       |           |
|    ToStringDashedHex | COMPlus_EnableSSE2=0 |  83.17 ns | 0.439 ns | 0.389 ns |  1.00 |      96 B |
|    ToStringDashedHex |                Empty |  45.70 ns | 0.534 ns | 0.500 ns |  0.55 |      96 B |
|                      |                      |           |          |          |       |           |
| ToStringDashBrackets | COMPlus_EnableSSE2=0 |  86.79 ns | 1.672 ns | 2.553 ns |  1.00 |     104 B |
| ToStringDashBrackets |                Empty |  49.35 ns | 0.937 ns | 0.877 ns |  0.56 |     104 B |
|                      |                      |           |          |          |       |           |
|          ToStringHex | COMPlus_EnableSSE2=0 |  84.31 ns | 0.565 ns | 0.501 ns |  1.00 |      88 B |
|          ToStringHex |                Empty |  37.47 ns | 0.539 ns | 0.504 ns |  0.44 |      88 B |